requirements.txt
flask==3.0.0
python-dotenv==1.0.0
langchain==0.1.20
langchain-groq==0.1.3
langchain-core==0.1.52